test: split test_btanalysis and test_metrics
